requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.5
django-environ==0.9.0
django-filter==22.1
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
executing==1.2.0
icecream==2.1.3
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
Markdown==3.4.1
model-bakery==1.10.1
Pillow==9.4.0
prompt-toolkit==3.0.36
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-decouple==3.7
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6

django-allauth==0.50.0
django-crispy-forms==1.14.0
drf-writable-nested==0.6.3
django-restql==0.15.3
django-multiselectfield==0.1.12
django-model-utils==4.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
django-versatileimagefield==2.2
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
PyMySQL==1.0.2
django-environ==0.9.0
pylint==2.14.4
pylint_django==2.5.3
Faker==13.15.1
factory-boy==3.2.1
mimesis==4.1.3
mimesis_factory==1.2.0
pytest==7.1.2
pytest_steps==1.8.0
pytest-sugar==0.9.5
pytest-django==4.5.2
firebase-admin==6.0.1
django-url-filter==0.3.15
djangorestframework-filters==1.0.0.dev2
django-cors-headers==3.13.0
channels==4.0.0
uvicorn[standard]==0.20.0
icecream
